feat: Always create macro
I think this rule means only evaluate at compile time so the while
macro was acting up again. It seems like eval-when is used to make sure
macros that rely on functions exist when the macro runs at compile time.

Since the while macro relied on standard Common Lisp, placing it in an
eval when was unnecessary